requirements.txt
fastapi[all]
uvicorn
email-validator
dateutils
## python-smbus
## adafruit-circuitpython-servokit
PCA9685-driver
